refactor(nexus): deleted tests
