--- crates/nexus/tests/common/mod.rs
@@ -1,3 +1,0 @@
-pub mod mock;
-pub mod read_side_setup;
-pub mod write_side_setup;
